+++ subsys/usb/common/CMakeLists.txt
@@ -1,0 +1,4 @@
+# Copyright (c) 2025 Nordic Semiconductor
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_include_directories(include)
