+++ subsys/usb/device_next/class/usbd_uvc.c
@@ -25,7 +25,8 @@
 #include <zephyr/usb/usb_ch9.h>
 #include <zephyr/usb/class/usbd_uvc.h>
 
-#include "usbd_uvc.h"
+#include "usb_uvc.h"
+
 #include "../../../drivers/video/video_ctrls.h"
 #include "../../../drivers/video/video_device.h"
